Moved to solar_system_location - HOYS_DEFAULT - EoT and d functions modified
