+++ CombiCSP/SolarGeometry.py
@@ -11,8 +11,10 @@ import pandas as pd
 # import pvlib
 # from pvlib import clearsky, atmosphere, solarposition
 
-# TODO mode HOYS_DEFAULT
-HOYS_DEFAULT = np.arange(1, 8761, 1) # hours of year
+from CombiCSP import HOYS_DEFAULT
+from CombiCSP.solar_system_location import EoT, d 
+
+
 
 
 #TODO Move this parameters to another location because now the system can only calculate for these long and lat.
@@ -22,107 +24,6 @@ mer = -25 # for Greece check to replace with 15 * dt_gmt
 lon = 24 # Crete 35.2401° N, 24.8093° E [east negative, west positive]
 
 
-#%% ===================================== earth declination angles
-def eda(hoy:np.array=HOYS_DEFAULT, method:str= 'wiki' ):
-    """wrapper function for the different earth declination angles functions
-
-    Useful reading [sunpos.py](https://levelup.gitconnected.com/python-sun-position-for-solar-energy-and-research-7a4ead801777)
-
-    Args:
-        hoy (np.array, optional): _description_. Defaults to HOYS_DEFAULT.
-        method (str, optional): _description_. Defaults to 'wiki'.
-
-    Returns:
-        _type_: _description_
-    """    
-    
-    dic = {'wiki':d,
-           'Katsaprakakis': d2,
-           '-81': d3,
-           'pveducation' :d1
-           }
-    if  method not in dic.keys():
-        raise(ValueError(f"method can be [{dic.keys()}]"))
-    return dic.get(method,None)(hoy)
-
-def d1(hoy:np.array=HOYS_DEFAULT): # earth declination angle [in degrees]
-    """earth declination angle 
-
-    The +10 comes from the fact that the winter solstice occurs before the start
-    of the year. The equation also assumes that the suns orbit is a perfect circle and 
-    the factor of 360/365 converts the day number to a position in the orbit.
-
-    Args:
-        hoy (np.array, optional): an array in hours of year . Defaults to HOYS_DEFAULT.
-
-    Returns:
-        _type_: earth inclination angle in degrees
-    """
-    # https://www.pveducation.org/pvcdrom/properties-of-sunlight/declination-angle#footnote1_osno74c'''
-    return -23.45 * np.cos( np.radians(360*(hoy+10*24)/365*24))
-
-def d2(hoy:np.array=HOYS_DEFAULT):
-    '''pp.6, Appendix C in D.A. Katsaprakakis, Power Plant Synthesis, CRC Press, 2020.
-    https://doi.org/10.1201/b22190.'''
-    return 23.45 * np.sin(np.radians(360*(hoy+284*24)/365*24))
-
-def d3(hoy:np.array=HOYS_DEFAULT):
-    return 23.45 * np.sin(np.radians(360*(hoy-81*24)/365*24))
-
-def d(hoy:np.array=HOYS_DEFAULT): # [in radians] https://en.wikipedia.org/wiki/Sunrise_equation
-    '''Calculate ecliptic coordinates (ecliptic longitude and obliquity of the
-    ecliptic in radians but without limiting the angle to be less than 2*Pi
-    (i.e., the result may be greater than 2*Pi)'''
-    dOmega = 2.1429 - 0.0010394594 * hoy
-    dMeanLongitude = 4.8950630 + 0.017202791698 * hoy
-    dMeanAnomaly = 6.2400600 + 0.0172019699 * hoy
-    dEclipticLongitude = dMeanLongitude + 0.03341607 * np.sin(dMeanAnomaly) 
-    + 0.00034894 * np.sin( 2 * dMeanAnomaly) - 0.0001134 - 0.0000203 * np.sin(dOmega)
-    dEclipticObliquity = 0.4090928 - 6.2140e-9 * hoy 
-    + 0.0000396 * np.cos(dOmega)
-    '''Calculate celestial coordinates ( right ascension and declination ) in radians
-    but without limiting the angle to be less than 2*Pi (i.e., the result may be
-    greater than 2*Pi)'''
-    dSin_EclipticLongitude = np.sin( dEclipticLongitude )
-    dY = np.cos( dEclipticObliquity ) * dSin_EclipticLongitude
-    dX = np.cos( dEclipticLongitude )
-    dRightAscension = np.arctan2( dY,dX )
-    if dRightAscension.any() < 0: dRightAscension = dRightAscension + 2*np.pi
-    return np.arcsin( np.sin( dEclipticObliquity ) * dSin_EclipticLongitude )
-
-#%% Equations of time
-
-def EoT(hoy:np.array=HOYS_DEFAULT): # equation of time [in minutes]
-    gamma = 360*(hoy-1)/365
-    return 2.2918*(0.0075+0.1868*np.cos(np.radians(gamma))-3.2077*np.sin(np.radians(gamma)) \
-        -1.4615*np.cos(np.radians(2*gamma))-4.089*np.sin(np.radians(2*gamma)))
-
-# TODO this do not work due to dayofyear --- uncomment when this is clear.
-# def _calculate_simple_day_angle(dayofyear, offset=1):
-#     """simple method for calculating the solar angle
-
-#     Args:
-#         dayofyear (_type_): _description_
-#         offset (int, optional): _description_. Defaults to 1.
-
-#     Returns:
-#         _type_: solar angle  in radians 
-#     """    
-#     return (2. * np.pi / 365.) * (dayofyear - offset)
-
-# def EoTS(hoy): # Equation of time from Duffie & Beckman and attributed to Spencer
-#     #(1971) and Iqbal (1983) [in minutes]
-#     day_angle = _calculate_simple_day_angle(dayofyear)
-#     return (1440.0 / 2 / np.pi) * (0.0000075 +
-#     0.001868 * np.cos(day_angle) - 0.032077 * np.sin(day_angle) -
-#     0.014615 * np.cos(2.0 * day_angle) - 0.040849 * np.sin(2.0 * day_angle))
-
-# def EoTPVCDROM(hoy): # equation of time [in minutes]
-#     # PVCDROM: http://www.pveducation.org/pvcdrom/2-properties-sunlight/solar-time
-#     # Soteris A. Kalogirou, "Solar Energy Engineering Processes and
-#     # Systems, 2nd Edition" Elselvier/Academic Press (2009).
-#     bday = _calculate_simple_day_angle(dayofyear) - (2.0 * np.pi / 365.0) * 80.0
-#     return 9.87 * np.sin(2.0 * bday) - 7.53 * np.cos(bday) - 1.5 * np.sin(bday)
 
 #%% 
 def tsol(hoy:np.array=HOYS_DEFAULT): # solar time [in decimal hours] introduce if function for east/west<<<<<<<<<
