+++ CombiCSP/__init__.py
@@ -4,8 +4,9 @@
     @Date: 2022/06/30
     @Credit: the functions were originally in  CSPEcon.py 
 """
-from CombiCSP.SolarGeometry import HOYS_DEFAULT
+from CombiCSP.solar_system_location import HOYS_DEFAULT
 from CombiCSP.misc import CtoK, OutputContainer
 from CombiCSP.solar_system_location import SolarSystemLocation
 from CombiCSP.solar_tower import SolarTowerCalcs
 from CombiCSP.solar_trough import SolarTroughCalcs
+
